Class "STBP". Result of a Sequential tests of Bayesian posterior probabilities

This class encapsulates results of a Sequential tests of Bayesian posterior probabilities

Slots

call

(language) The call to stbp_simple or stbp_composite.

probabilities

(numeric) Vector of sequential posterior probabilities.

recommendation

(character) Recommendation on H, whether to accept, reject or keep sampling.

iterations

(numeric) Number of sequential sampling bouts required or processed.

Examples

set.seed(101)
counts3 <- rpois(5, lambda = 3)

test1F <- stbp_composite(data = counts3,
                          greater_than = TRUE,
                          hypothesis = 5,
                          density_func = "poisson",
                          prior = 0.5,
                          lower_bnd = 0,
                          upper_bnd = Inf,
                          lower_criterion = 0.001,
                          upper_criterion = 0.999)
test1F # returns "reject H".
#> 
#> Sequential test of Bayesian posterior probabilities
#> Family: poisson
#> H: mu > 5
#> Probability: 0.00097 from 2 sampling bouts
#> Recommendation based on provided criteria: reject H

counts10 <- matrix(rep(0, 30), 10, 3)

test1G <- stbp_simple(data = counts10,
                        density_func= "poisson",
                        prior = 0.5,
                        upper_bnd = Inf,
                        lower_criterion = 0,
                        upper_criterion = 0.9999)
test1G # returns "keep sampling" due to insufficient evidence.
#> 
#> Sequential test of Bayesian posterior probabilities
#> Family: poisson
#> H: mu = 0
#> Probability: 0.999 from 3 sampling bouts
#> Recommendation based on provided criteria: keep sampling

## End (Not run)